+++ bloc_signals/lib/bloc_signals.dart
@@ -12,4 +12,5 @@ export 'src/concurrency/event_transformers.dart';
 export 'src/concurrency/mutex.dart';
 export 'src/devtools_observer.dart';
 export 'src/devtools_service.dart';
+export 'src/signal_adapter.dart';
 export 'src/stream_adapter.dart';